final script

+++ root_macros/printTables.sh
@@ -8,7 +8,10 @@ fi
 Input=$1
 Output=`pwd`/$2
 
-mv -f $Output{,.bck} 
+if [ -f $Output ]
+then
+    mv -f $Output{,.bck} 
+fi
 
 root -b -q 'printNEvts.C+('\"$Input\"', -1)' >> $Output
 root -b -q 'printNEvts.C+('\"$Input\"', 0)' >> $Output
@@ -19,10 +22,12 @@ root -b -q 'printNEvts.C+('\"$Input\"', 3)' >> $Output
 cd ../Limits
 
 #Breakdown of Bkg After Ht Cut
-root -b -q 'ExpectedEvts.C+('\"$Input\"',"cutValues.wz.dat",-1, "scaleMC tbl noWind")' >> $Output
+#Don't scale MC for this table
+root -b -q 'ExpectedEvts.C+('\"$Input\"',"cutValues.wz.dat",-1, "tbl noWind")' >> $Output
 root -b -q 'printTable.C+(1)' >> $Output
 
 #Breakdown of Bkg After Window Cut
-root -b -q 'ExpectedEvts.C+('\"$Input\"',"cutValues.wz.dat",-1, "scaleMC tbl")' >> $Output
+root -b -q 'ExpectedEvts.C+('\"$Input\"',"cutValues.wz.dat",-1, "tbl")' >> $Output
+root -b -q 'ExpectedEvts.C+('\"$Input\"',"cutValues.wz.dat",-1, "scaleMC tbl")' >> junk.txt
 root -b -q 'printTable.C+(0)' >> $Output
